redirect to login page if the user not logged in and add move images to assets folder
