+++ src/assets/img/shutUpLogo.png
